Make migrations work in Laravel 6

+++ migrations/2018_09_30_110932_create_forms_table.php
@@ -21,7 +21,7 @@ class CreateFormsTable extends Migration
         Schema::create('forms', function (Blueprint $table) {
             $table->increments('id');
 
-            $table->unsignedInteger('user_id')->nullable();
+            $table->unsignedBigInteger('user_id')->nullable();
 
             $table->string('name');
             $table->string('visibility');
